Adding support for Windows-style backslashes
By replacing backslashes by forward slashes and simple features of scan and verify bring add support for the backslash in directory names.
The new code works fine on Cygwin and plain Windows, but I get a build error on trim_sep on MinGW. I have to sort this one out. It may be a problem with the support of submodules in gfortran.